Refactored MatchGrid.tsx
