+++ components/gameProgress/MatchesGrid/index.ts
@@ -1,0 +1,2 @@
+import MatchesGridContainer from "./MatchesGridContainer";
+export default MatchesGridContainer;
